executor: move tests
